oct-1907: allocation drawer
