+++ client/src/components/Allocation/Allocation.module.scss
@@ -1,0 +1,56 @@
+.root {
+  padding-bottom: 8.8rem;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+
+  @media #{$tablet-up} {
+    padding-bottom: 14.4rem;
+  }
+
+  @media #{$desktop-up} {
+    padding-bottom: 0;
+    width: 68rem;
+    min-height: 0;
+  }
+
+  .title {
+    padding: 0;
+    display: flex;
+    align-items: center;
+    min-height: 8.8rem;
+    width: 100%;
+    color: $color-octant-dark;
+    font-size: $font-size-24;
+    font-weight: $font-weight-bold;
+    line-height: 1.4rem;
+
+    @media #{$tablet-up} {
+      font-size: $font-size-32;
+      min-height: 14.4rem;
+    }
+
+    @media #{$desktop-up} {
+      padding: 0 4rem;
+    }
+  }
+
+  .box {
+    &:not(:last-child) {
+      margin: 0 auto 1.6rem;
+    }
+  }
+
+  .boxesWrapper {
+    width: 100%;
+
+    &.withMarginBottom {
+      margin-bottom: 12.8rem;
+    }
+
+    @media #{$desktop-up} {
+      padding: 0 4rem 4rem;
+      overflow: auto;
+    }
+  }
+}
