+++ client/src/components/Allocation/AllocationNavigation/types.ts
@@ -1,12 +1,8 @@
-import { CurrentView } from 'views/AllocationView/types';
-
 export default interface AllocationNavigationProps {
   areButtonsDisabled: boolean;
-  currentView: CurrentView;
   isLeftButtonDisabled: boolean;
   isLoading: boolean;
   isWaitingForAllMultisigSignatures?: boolean;
   onAllocate: () => void;
   onResetValues: () => void;
-  setCurrentView: (newView: CurrentView) => void;
 }
