+++ client/src/components/Allocation/AllocationRewardsBox/AllocationRewardsBox.module.scss
@@ -1,106 +1,33 @@
 .root {
-  position: relative;
-  color: $color-octant-grey5;
-  padding: 2rem 2.4rem 2.3rem;
+  min-height: 7.2rem;
+  max-height: 7.2rem;
+  border-radius: $border-radius-16;
+  background-color: $color-octant-grey8;
+  padding: 1.6rem 1.6rem 0;
+  font-weight: $font-weight-bold;
+  text-align: left;
+  margin-bottom: 1.6rem;
 
-  @media #{$tablet-down} {
-    padding: 1.6rem;
-  }
-}
-
-.title {
-  font-size: $font-size-18;
-  margin: 0 0 0 0.8rem;
-  color: $color-octant-green;
-  line-height: 2.4rem;
-
-  &.greyTitle {
-    color: $color-octant-grey5;
-  }
-
-  @media #{$tablet-down} {
-    font-size: $font-size-16;
-    line-height: 2rem;
-  }
-}
-
-.subtitle {
-  margin-left: 0.8rem;
-}
-
-.isManualBadge {
-  position: absolute;
-  top: 2.4rem;
-  right: 2.4rem;
-  border: 0.1rem solid $color-octant-green;
-  padding: 0.2rem 0.5rem;
-  text-transform: uppercase;
-  border-radius: $border-radius-04;
-  color: $color-octant-green;
-  background: $color-octant-green5;
-  line-height: 1.4rem;
-}
-
-.sliderWrapper {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 7.2rem;
-  background-color: $color-octant-grey6;
-  border-radius: $border-radius-08;
-  padding: 0 1.6rem;
-  margin: 1.6rem 0 1.2rem;
-
-  &.isManuallyEdited {
-    background-color: $color-octant-green5;
-  }
-
-  @media #{$tablet-down} {
-    margin: 1.1rem 0 0.8rem;
-  }
-}
-
-.sections {
-  display: flex;
-  justify-content: space-between;
-  align-self: stretch;
-  margin: 0 0.8rem;
-}
-
-.section {
-  &:not(.isDisabled):not(.isLocked) {
-    cursor: pointer;
-  }
-
-  .header {
-    line-height: 1.6rem;
-    padding: 0.4rem 0 0.1rem 0;
+  @media #{$tablet-up} {
+    padding: 2rem 2.4rem 0;
   }
 
   .value {
-    line-height: 2rem;
-    font-size: $font-size-18;
-    color: $color-octant-dark;
-    height: 1.6rem;
+    color: $color-octant-green;
+    font-size: $font-size-16;
+    line-height: 2.4rem;
+
+    @media #{$tablet-up} {
+      font-size: $font-size-18;
+    }
 
     &.isGrey {
       color: $color-octant-grey5;
     }
-
-    @media #{$tablet-down} {
-      font-size: $font-size-16;
-    }
   }
 
-  &:first-child {
-    text-align: left;
-  }
-
-  &:last-child {
-    text-align: right;
-  }
-
-  &.isDisabled {
+  .label {
+    font-size: $font-size-12;
     color: $color-octant-grey5;
   }
 }
