+++ client/src/components/Allocation/AllocationRewardsBox/types.ts
@@ -1,8 +1,4 @@
 export default interface AllocationRewardsBoxProps {
-  className?: string;
   isDisabled?: boolean;
-  isError: boolean;
   isLocked?: boolean;
-  isManuallyEdited?: boolean;
-  setRewardsForProjectsCallback: ({ rewardsForProjectsNew }) => void;
 }
