+++ client/src/components/Allocation/AllocationSliderBox/AllocationSliderBox.module.scss
@@ -1,0 +1,87 @@
+.root {
+  position: relative;
+  color: $color-octant-grey5;
+  padding: 1.6rem;
+
+  @media #{$tablet-up} {
+    padding: 2.4rem;
+  }
+}
+
+.isManualBadge {
+  position: absolute;
+  top: 2.4rem;
+  right: 2.4rem;
+  border: 0.1rem solid $color-octant-green;
+  padding: 0.2rem 0.5rem;
+  text-transform: uppercase;
+  border-radius: $border-radius-04;
+  color: $color-octant-green;
+  background: $color-octant-green5;
+  line-height: 1.4rem;
+}
+
+.sliderWrapper {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  background-color: $color-octant-grey6;
+  border-radius: $border-radius-08;
+  padding: 0 1.6rem;
+  margin-bottom: 0.8rem;
+  height: 7.2rem;
+
+  &.isManuallyEdited {
+    background-color: $color-octant-green5;
+  }
+
+  @media #{$tablet-up} {
+    margin-bottom: 1.6rem;
+    height: 10.4rem;
+  }
+}
+
+.sections {
+  display: flex;
+  justify-content: space-between;
+  align-self: stretch;
+  margin: 0 0.8rem;
+}
+
+.section {
+  &:not(.isDisabled):not(.isLocked) {
+    cursor: pointer;
+  }
+
+  .header {
+    font-size: $font-size-12;
+    line-height: 2.4rem;
+  }
+
+  .value {
+    line-height: 2rem;
+    font-size: $font-size-16;
+    color: $color-octant-dark;
+    height: 1.6rem;
+
+    &.isGrey {
+      color: $color-octant-grey5;
+    }
+
+    @media #{$tablet-up} {
+      font-size: $font-size-18;
+    }
+  }
+
+  &:first-child {
+    text-align: left;
+  }
+
+  &:last-child {
+    text-align: right;
+  }
+
+  &.isDisabled {
+    color: $color-octant-grey5;
+  }
+}
