+++ client/src/components/Allocation/AllocationSliderBox/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './AllocationSliderBox';
