+++ client/src/components/Allocation/AllocationSliderBox/types.ts
@@ -1,0 +1,8 @@
+export default interface AllocationSliderBoxProps {
+  className?: string;
+  isDisabled?: boolean;
+  isError: boolean;
+  isLocked?: boolean;
+  isManuallyEdited?: boolean;
+  setRewardsForProjectsCallback: ({ rewardsForProjectsNew }) => void;
+}
