+++ client/src/components/Allocation/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './Allocation';
