+++ client/src/constants/domElementsIds.ts
@@ -2,4 +2,5 @@ export const METRICS_EPOCH_ID = 'MetricsEpoch';
 export const METRICS_GENERAL_ID = 'MetricsGeneral';
 export const METRICS_PERSONAL_ID = 'MetricsPersonal';
 export const LAYOUT_BODY_ID = 'LayoutBody';
+export const LAYOUT_NAVBAR_ID = 'LayoutNavbar';
 export const EPOCH_RESULTS_BAR_ID = 'EpochResultsBar';
