+++ client/src/locales/en/translation.json
@@ -55,6 +55,49 @@
     "waitingForConfirmation": "Waiting for confirmation"
   },
   "components": {
+    "allocation": {
+      "allocateRewards": "Allocate rewards",
+      "lowUQScoreModal": {
+        "header": "Your UQ score is too low!",
+        "text": "If you allocate with this UQ score, you will receive <0>80% less match funding</0> than if you are verified. You cannot change this after allocation. Are you sure you want to proceed?",
+        "checkboxLabel": "I understand I will receive 20% of the maximum match funding",
+        "goToUQSettings": "Go to UQ Settings",
+        "proceedToAllocate": "Proceed to allocate"
+      },
+      "multisigSignatureToast": {
+        "title": "Multi-sig: signature needed",
+        "message": "At least one signature is required to start allocation. Please do not close Octant."
+      },
+      "allocationItem": {
+        "epoch1": "No funding threshold in Epoch 1",
+        "standard": "{{sum}} / {{threshold}}",
+        "simulate": "Your impact {{value}}",
+        "simulateLoading": "Calculating..."
+      },
+      "allocationSuccessful": "Allocation successful",
+      "tip": {
+        "rewards": {
+          "title": "Time to allocate",
+          "text": {
+            "desktop": "Add projects and use the slider to divide rewards between yourself and your chosen projects",
+            "mobile": "Add projects and use the slider to divide rewards between yourself & your chosen projects"
+          }
+        },
+        "uqTooLow": {
+          "title": {
+            "desktop": "Your uniqueness score is too low",
+            "mobile": "Your UQ score is too low"
+          },
+          "text": {
+            "desktop": "You will receive only 20% of the maximum match funding. Visit <0>Settings</0> to recalculate it before you allocate and improve your match funding",
+            "mobile": "You will receive only 20% of the max. match funding. Visit <0>Settings</0> to recalculate before you allocate"
+          }
+        }
+      },
+      "modalAllocationValuesEdit": {
+        "header": "Edit {{allocation}}"
+      }
+    },
     "shared": {
       "onboardingStepper": {
         "reopenOnboarding": "Reopen onboarding"
@@ -428,7 +471,7 @@
         "docs": "Docs",
         "brandAssets": "Brand assets",
         "termsOfUse": "Terms of use",
-        "farcaster":  "Farcaster",
+        "farcaster": "Farcaster",
         "twitterX": "Twitter/X"
       }
     },
@@ -458,48 +501,6 @@
   "views": {
     "home": {
       "title": "Welcome to Epoch {{epoch}}"
-    },
-    "allocation": {
-      "lowUQScoreModal": {
-        "header": "Your UQ score is too low!",
-        "text": "If you allocate with this UQ score, you will receive <0>80% less match funding</0> than if you are verified. You cannot change this after allocation. Are you sure you want to proceed?",
-        "checkboxLabel": "I understand I will receive 20% of the maximum match funding",
-        "goToUQSettings": "Go to UQ Settings",
-        "proceedToAllocate": "Proceed to allocate"
-      },
-      "multisigSignatureToast": {
-        "title": "Multi-sig: signature needed",
-        "message": "At least one signature is required to start allocation. Please do not close Octant."
-      },
-      "allocationItem": {
-        "epoch1": "No funding threshold in Epoch 1",
-        "standard": "{{sum}} / {{threshold}}",
-        "simulate": "Your impact {{value}}",
-        "simulateLoading": "Calculating..."
-      },
-      "allocationSuccessful": "Allocation successful",
-      "tip": {
-        "rewards": {
-          "title": "Time to allocate",
-          "text": {
-            "desktop": "Add projects and use the slider to divide rewards between yourself and your chosen projects",
-            "mobile": "Add projects and use the slider to divide rewards between yourself & your chosen projects"
-          }
-        },
-        "uqTooLow": {
-          "title": {
-            "desktop": "Your uniqueness score is too low",
-            "mobile": "Your UQ score is too low"
-          },
-          "text": {
-            "desktop": "You will receive only 20% of the maximum match funding. Visit <0>Settings</0> to recalculate it before you allocate and improve your match funding",
-            "mobile": "You will receive only 20% of the max. match funding. Visit <0>Settings</0> to recalculate before you allocate"
-          }
-        }
-      },
-      "modalAllocationValuesEdit": {
-        "header": "Edit {{allocation}}"
-      }
     },
     "earn": {
       "tips": {
