+++ client/src/store/allocations/types.ts
@@ -1,5 +1,8 @@
+export type CurrentView = 'edit' | 'summary';
+
 export interface AllocationsData {
   allocations: string[];
+  currentView: CurrentView;
   rewardsForProjects: bigint;
 }
 
@@ -7,5 +10,6 @@ export interface AllocationsMethods {
   addAllocations: (projectAddresses: string[]) => void;
   removeAllocations: (projectAddresses: string[]) => void;
   setAllocations: (projectAddresses: string[]) => void;
+  setCurrentView: (currentView: CurrentView) => void;
   setRewardsForProjects: (rewardsForProjects: bigint) => void;
 }
