+++ client/src/store/layout/types.ts
@@ -1,12 +1,12 @@
 export interface LayoutData {
-  showAllocateDrawer: boolean;
+  showAllocationDrawer: boolean;
   showConnectWalletModal: boolean;
   showSettingsDrawer: boolean;
   showWalletModal: boolean;
 }
 
 export interface LayoutMethods {
-  setShowAllocateDrawer: (showAllocateDrawer: boolean) => void;
+  setShowAllocationDrawer: (showAllocationDrawer: boolean) => void;
   setShowConnectWalletModal: (showConnectWalletModal: boolean) => void;
   setShowSettingsDrawer: (showSettingsDrawer: boolean) => void;
   setShowWalletModal: (showWalletModal: boolean) => void;
